commit 12/4/20
historical data consolidation/summaries, and other minor script edits
