+++ Python/Deprecated/analyze_developable_area.py
