+++ Python/prep_historical_aggregation.py
@@ -27,6 +27,72 @@ def prep_historical_aggregation(aggregating_geometry_path,
                                 lu_reference_path = None,
                                 lu_reference_field = None):
     
+    '''
+    Parameters
+    ----------
+    aggregating_geometry_path: path
+        file path to polygon feature class, to which inputs will be aggregated
+        for reporting
+    aggregating_geometry_id_field: str
+        field name of unique identifier for aggregating geometry features
+    years: list of ints
+        years of historical data to be aggregated
+    inputs_gdb_format: directory
+        gdb (/FeatureClass) location of the inputs, with a formatting wild
+        card for year; see notes on how formatting should be done
+    inputs_fc_format: str
+        name of inputs feature class within `inputs_gdb_format`, with a
+        formatting wild card for year; see notes on how formatting should
+        be done
+    inputs_field_names: list of str
+        field names of `inputs` attributes to be aggregated
+    save_gdb_path: path
+        file path for saving the aggregated historical data
+    lu_field: str
+        land use field in the inputs; should only be provided if the `inputs
+        fields` should be summarized by land use as well as `aggregating
+        geometry` feature. See notes for usage
+    lu_reference_path: path
+        path to a land use reference .csv; should only be provided if the
+        reference .csv contains the land use field on which you want to
+        aggregate, in which case the inputs will be joined to this table
+        on `lu_field`. See notes for usage
+    lu_reference_field: str
+        field name in the land use reference .csv defining the land use
+        attribute on which you want to aggregated. See notes for usage
+        
+    Notes
+    -----
+    1. For `inputs_gdb_format` and `inputs_fc_format`, the assumption is that
+    there is some consistent naming structure that points to the same data
+    over different years, and that this naming structure varies only on year.
+    Wherever a year appears in either the gdb directory or the feature class
+    name, replace the year with `{Year}` in the provided string. For example,
+    if data for 2010 is stored at `Data_2010.gdb/Features` (and similarly
+    for other years), then `inputs_gdb_format = 'Data_{Year}.gdb'` and 
+    `inputs_fc_format = 'Features'`. Similarly, if the data for 2010 is stored
+    at 'Data.gdb/Features_2010' (and similarly for other years), then
+    `inputs_gdb_format = 'Data.gdb'` and `inputs_fc_format = 'Features_{Year}'`
+    
+    2. The `lu_{}` fields need only be specified if you want the data
+    aggregated by land use as well as the aggregating geometry. `lu_field`
+    defines the land use field in the inputs. `lu_reference_path` should
+    only be provided if there is a reference table with a more appropriate
+    land use field on which to aggregate. If provided, it MUST contain
+    `lu_field` as a field, to provide a valid join field to the inputs; if not
+    provided but `lu_field` is, land use aggregation will be completed on
+    `lu_field` in the inputs. Finally, `lu_reference_field` is the name of the 
+    more appropriate land use field mentioned above, and must be provided if
+    `lu_reference_path` is.
+    
+    Returns
+    -------
+    Saves a feature class to `save_gdb_path` of `inputs_field_names`
+    aggregated to the geometries of `aggregating_geometry_path` (and
+    potentially a land use field) for all years in `years`. The save location
+    will be returned to confirm saving.
+    '''
+    
     # Initialize a template for results ---------------------------------
     
     print("")
@@ -70,6 +136,7 @@ def prep_historical_aggregation(aggregating_geometry_path,
     # outside of a .gdb, we run the risk of having our field names truncated,
     # which is bad if we want to keep referencing our input fields (which
     # we most definitely do)
+    print("-- setting up a temporary location for intermediates")
     temp_dir = tempfile.mkdtemp()
     arcpy.CreateFileGDB_management(out_folder_path = temp_dir,
                                    out_name = "Intermediates.gdb")
@@ -78,6 +145,11 @@ def prep_historical_aggregation(aggregating_geometry_path,
     # For loop processing, we'll store the results of each iteration in a list
     # that we'll later concat into a single df
     agdf = []
+    
+    # We're also relying on list comprehension of the input fields, so we'll
+    # verify that our input fields are in a list
+    if type(inputs_field_names) is not list:
+        inputs_field_names = [f for f in [inputs_field_names]]
     
     # Now, we're ready to complete the consolidation of our historical data.
     # Processing is going to operate in a loop. For each year, we're going
@@ -89,7 +161,7 @@ def prep_historical_aggregation(aggregating_geometry_path,
     
     for yr in years:
         syr = str(yr)
-        print("-- " + str(syr))
+        print("-- processing " + str(syr))
         
         # 1. read fields of interest
         print("---- reading fields of interest")
@@ -187,7 +259,7 @@ def prep_historical_aggregation(aggregating_geometry_path,
     # repetitions needed varies by feature. So, our first step is to
     # attribute the template feature class with a variable for the number
     # of repetitions
-    print("-- attributing template with repetition count")
+    print("-- attributing the template with a repetition count")
     
     template = arcpy.da.FeatureClassToNumPyArray(in_table = template_path,
                                                  field_names = aggregating_geometry_id_field,
@@ -225,16 +297,22 @@ def prep_historical_aggregation(aggregating_geometry_path,
                                         spatial_reference = sr)
     repeated_path = os.path.join(gdb_path, 
                                  "repeated")
-    
-    # Im here. Read in agdf from Downloads
-    with arcpy.da.SearchCursor(template_path, '*') as curs_in:
+    rep_fields = ["GEOID10","Count","SHAPE@"]
+    with arcpy.da.SearchCursor(template_path, rep_fields) as curs_in:
         flds_in = curs_in.fields
         idx_cnt = flds_in.index("Count")
-        with arcpy.da.InsertCursor(repeated_path, '*') as curs_out:
+        with arcpy.da.InsertCursor(repeated_path, rep_fields) as curs_out:
             for row in curs_in:
                 cnt = row[idx_cnt]
                 for i in range(0, cnt):
                     curs_out.insertRow(row)
+    
+    # We'll delete some intermediates here too
+    arcpy.DeleteField_management(repeated_path,
+                                  drop_field = "Count")
+    arcpy.Delete_management(template_path)
+    del curs_in
+    del curs_out
     
     # Next, we need to add a field to join on. Note that because our
     # initialized feature class and our results now have the same
@@ -243,42 +321,43 @@ def prep_historical_aggregation(aggregating_geometry_path,
     # attribute each with a sequential ID for easy binding. First, we'll add
     # the join field to the save feature class by sorting and calculating
     # a new field
+    print("-- attributing the save feature class with a join field")
     
-    
+    arcpy.Sort_management(in_dataset = repeated_path,
+                          out_dataset = save_gdb_path,
+                          sort_field = aggregating_geometry_id_field)
+    arcpy.Delete_management(repeated_path)
+    codeblock = 'val = 0 \ndef processID(): \n    global val \n    start = 1 \n    if (val == 0):  \n        val = start\n    else:  \n        val += 1  \n    return val'
     arcpy.AddField_management(in_table = save_gdb_path,
-                              field_name = "RepID",
-                              field_type = "SHORT")
+                              field_name = "JoinID",
+                              field_type = "LONG",
+                              field_is_required = "NON_REQUIRED")
+    arcpy.CalculateField_management(in_table = save_gdb_path,
+                                    field = "JoinID",
+                                    expression = "processID()",
+                                    expression_type = "PYTHON3",
+                                    code_block = codeblock)
     
-             
-    # Now our feature class and results data match, but we can't do a join 
-    # because our 'aggregating_geometry_id_field' is not unique (because of 
-    # repetitions). So, we need a repetition ID in both our data and the save 
-    # feature class for unique joining
-    print("-- attributing the results with a repetition ID")
+    # Now we do the same thing to our results
+    print("-- attributing the results with a join field")
     
-    rep_id = [np.arange(1, n+1).tolist() for n in counts["Count"]]
-    rep_id = [i for sublist in rep_id for i in sublist]
-    agid = np.repeat(counts[aggregating_geometry_id_field], counts["Count"]).tolist()
-    rep_id = pd.DataFrame.from_dict({aggregating_geometry_id_field: agid,
-                                     "RepID": rep_id})
     agdf = agdf.sort_values(aggregating_geometry_id_field).reset_index(drop=True)
-    rep_id = rep_id.sort_values([aggregating_geometry_id_field, "RepID"]).reset_index(drop=True)
-    agdf["RepID"] = rep_id["RepID"]
+    agdf["JoinID"] = [x for x in np.arange(1, len(agdf.index)+1)]
+    agdf = agdf.drop(columns = aggregating_geometry_id_field)
     
-    # Writing ----------------------------------------------------------------
+    # With both now atttributed with a JoinID, all that's left to do is join
+    # up our results
+    print("-- joining results to the save feature class")
     
-    print("")
-    print("Joining results to the initialized feature class")
-    
-    # The last step is then just joining our consolidated data to our
-    # initialized feature class  
-    df_et = np.rec.fromrecords(recList = all_agg.values, 
-                               names = all_agg.dtypes.index.tolist())
+    df_et = np.rec.fromrecords(recList = agdf.values, 
+                               names = agdf.dtypes.index.tolist())
     df_et = np.array(df_et)
-    arcpy.da.ExtendTable(in_table = ifc_path,
-                         table_match_field = aggregating_geometry_id_field,
+    arcpy.da.ExtendTable(in_table = save_gdb_path,
+                         table_match_field = "JoinID",
                          in_array = df_et,
-                         array_match_field = aggregating_geometry_id_field)
+                         array_match_field = "JoinID")
+    arcpy.DeleteField_management(in_table = save_gdb_path,
+                                 drop_field = "JoinID")
     
     # Done -------------------------------------------------------------------
     
@@ -292,17 +371,58 @@ def prep_historical_aggregation(aggregating_geometry_path,
 
 # This is for parcel characteristics (e.g. living area)
 if __name__ == "__main__":
-    # Inputs
+    # Inputs for TOTAL LIVING AREA AND PARCEL LAND AREA
     aggregating_geometry_path = "K:/Projects/MiamiDade/PMT/Data/Cleaned/Blocks.gdb/Blocks_2019"
     aggregating_geometry_id_field = "GEOID10"
     years = [2014, 2015, 2016, 2017, 2018, 2019]
     inputs_gdb_format = "K:/Projects/MiamiDade/PMT/Data/Cleaned/parcels.gdb"
     inputs_fc_format = "Miami_{Year}"
-    inputs_field_names = ["TOT_LVG_AREA", "LND_SQFOOT", "NO_RES_UNTS"]
-    save_gdb_path = "K:/Projects/MiamiDade/PMT/PMT_Trend.gdb/parcels/historical_parcel_chars"
+    inputs_field_names = ["TOT_LVG_AREA", "LND_SQFOOT"]
+    save_gdb_path = "K:/Projects/MiamiDade/PMT/Data/PMT_Trend.gdb/parcels/historical_FAR_inputs"
     lu_field = "DOR_UC"
     lu_reference_path = "K:/Projects/MiamiDade/PMT/Data/Reference/Land_Use_Recode.csv"
     lu_reference_field = "GN_VA_LU"
+    
+    # Inputs for NUMBER OF RESIDENTIAL UNITS
+    aggregating_geometry_path = "K:/Projects/MiamiDade/PMT/Data/Cleaned/Blocks.gdb/Blocks_2019"
+    aggregating_geometry_id_field = "GEOID10"
+    years = [2014, 2015, 2016, 2017, 2018, 2019]
+    inputs_gdb_format = "K:/Projects/MiamiDade/PMT/Data/Cleaned/parcels.gdb"
+    inputs_fc_format = "Miami_{Year}"
+    inputs_field_names = "NO_RES_UNTS"
+    save_gdb_path = "K:/Projects/MiamiDade/PMT/Data/PMT_Trend.gdb/parcels/historical_res_units"
+    lu_field = None
+    lu_reference_path = None
+    lu_reference_field = None
+    
+    # Inputs for TOTAL JOBS AND POPULATION
+    aggregating_geometry_path = "K:/Projects/MiamiDade/PMT/Data/Cleaned/Blocks.gdb/Blocks_2019"
+    aggregating_geometry_id_field = "GEOID10"
+    years = [2014, 2015, 2016, 2017, 2018, 2019]
+    inputs_gdb_format = "K:/Projects/MiamiDade/PMT/PMT_{Year}.gdb/Parcels"
+    inputs_fc_format = "socioeconomic_and_demographic"
+    inputs_field_names = ["Total_Employment", "Total_Population"]
+    save_gdb_path = "K:/Projects/MiamiDade/PMT/Data/PMT_Trend.gdb/parcels/historical_total_jobs_pop"
+    lu_field = None
+    lu_reference_path = None
+    lu_reference_field = None
+    
+    # Inputs for COMMUTES
+    aggregating_geometry_path = "K:/Projects/MiamiDade/PMT/Data/Cleaned/Blocks.gdb/Blocks_2019"
+    aggregating_geometry_id_field = "GEOID10"
+    years = [2014, 2015, 2016, 2017, 2018, 2019]
+    inputs_gdb_format = "K:/Projects/MiamiDade/PMT/PMT_{Year}.gdb/Parcels"
+    inputs_fc_format = "socioeconomic_and_demographic"
+    inputs_field_names = ["Total_Commutes",
+                          "Drove_PAR","Carpool_PAR","Transit_PAR",
+                          "NonMotor_PAR","Work_From_Home_PAR","AllOther_PAR"]
+    save_gdb_path = "K:/Projects/MiamiDade/PMT/Data/PMT_Trend.gdb/parcels/historical_commutes"
+    lu_field = None
+    lu_reference_path = None
+    lu_reference_field = None
+    
+    # Inputs for JOBS BY INDUSTRY CLUSTER
+    
     
     # Function
     prep_historical_aggregation(aggregating_geometry_path = aggregating_geometry_path,
